refactor: move shortlink plugin to _11ty
